+++ console/src/components/Services/Data/DataSources/ConnectionSettingsForm.tsx
@@ -6,6 +6,7 @@ import Tooltip from '../../../Common/Tooltip/Tooltip';
 import { getSupportedDrivers } from '../../../../dataSources';
 
 import styles from './DataSources.scss';
+import { IsolationLevelOptions } from '../../../../metadata/types';
 
 export interface ConnectionSettingsFormProps {
   // Connect DB State Props
@@ -50,6 +51,7 @@ const ConnectionSettingsForm: React.FC<ConnectionSettingsFormProps> = ({
               <div className={styles.connection_settings_input_layout}>
                 <LabeledInput
                   label="Max Connections"
+                  tooltipText="Maximum number of connections to be kept in the pool"
                   type="number"
                   className={`form-control ${styles.connnection_settings_form_input}`}
                   placeholder="50"
@@ -71,6 +73,7 @@ const ConnectionSettingsForm: React.FC<ConnectionSettingsFormProps> = ({
               <div className={styles.connection_settings_input_layout}>
                 <LabeledInput
                   label="Idle Timeout"
+                  tooltipText="The idle timeout (in seconds) per connection"
                   type="number"
                   className={`form-control ${styles.connnection_settings_form_input}`}
                   placeholder="180"
@@ -95,6 +98,7 @@ const ConnectionSettingsForm: React.FC<ConnectionSettingsFormProps> = ({
                 <div className={styles.connection_settings_input_layout}>
                   <LabeledInput
                     label="Retries"
+                    tooltipText="Number of retries to perform"
                     type="number"
                     className={`form-control ${styles.connnection_settings_form_input}`}
                     placeholder="1"
@@ -113,6 +117,82 @@ const ConnectionSettingsForm: React.FC<ConnectionSettingsFormProps> = ({
                   />
                 </div>
               ) : null}
+              {getSupportedDrivers('connectDbForm.pool_timeout').includes(
+                connectionDBState.dbType
+              ) ? (
+                <div className={styles.connection_settings_input_layout}>
+                  <LabeledInput
+                    label="Pool Timeout"
+                    tooltipText="Maximum time (in seconds) to wait while acquiring a Postgres connection from the pool"
+                    type="number"
+                    className={`form-control ${styles.connnection_settings_form_input}`}
+                    placeholder="360"
+                    value={
+                      connectionDBState.connectionSettings?.pool_timeout ??
+                      undefined
+                    }
+                    onChange={e =>
+                      connectionDBStateDispatch({
+                        type: 'UPDATE_POOL_TIMEOUT',
+                        data: e.target.value,
+                      })
+                    }
+                    min="0"
+                    boldlabel
+                    data-test="pool-timeout"
+                  />
+                </div>
+              ) : null}
+              {getSupportedDrivers(
+                'connectDbForm.connection_lifetime'
+              ).includes(connectionDBState.dbType) ? (
+                <div className={styles.connection_settings_input_layout}>
+                  <LabeledInput
+                    label="Connection Lifetime"
+                    tooltipText="Time (in seconds) from connection creation after which the connection should be destroyed and a new one created. A value of 0 indicates we should never destroy an active connection. If 0 is passed, memory from large query results may not be reclaimed."
+                    type="number"
+                    className={`form-control ${styles.connnection_settings_form_input}`}
+                    placeholder="600"
+                    value={
+                      connectionDBState.connectionSettings
+                        ?.connection_lifetime ?? undefined
+                    }
+                    onChange={e =>
+                      connectionDBStateDispatch({
+                        type: 'UPDATE_CONNECTION_LIFETIME',
+                        data: e.target.value,
+                      })
+                    }
+                    min="0"
+                    boldlabel
+                    data-test="connection-lifetime"
+                  />
+                </div>
+              ) : null}
+              {getSupportedDrivers('connectDbForm.isolation_level').includes(
+                connectionDBState.dbType
+              ) && (
+                <div className={styles.connection_settings_input_layout}>
+                  <label>
+                    <b>Isolation Level</b>
+                    <Tooltip message="The transaction isolation level in which the queries made to the source will be run" />
+                  </label>
+                  <select
+                    className={`form-control ${styles.connnection_settings_form_input}`}
+                    onChange={e =>
+                      connectionDBStateDispatch({
+                        type: 'UPDATE_ISOLATION_LEVEL',
+                        data: e.target.value as IsolationLevelOptions,
+                      })
+                    }
+                    value={connectionDBState.isolationLevel}
+                  >
+                    <option value="read-committed">read-committed</option>
+                    <option value="repeatable-read">repeatable-read</option>
+                    <option value="serializable">serializable</option>
+                  </select>
+                </div>
+              )}
               {getSupportedDrivers(
                 'connectDbForm.prepared_statements'
               ).includes(connectionDBState.dbType) ? (