+++ console/src/dataSources/services/bigquery/index.tsx
@@ -158,8 +158,11 @@ export const supportedFeatures: SupportedFeaturesType = {
     environmentVariable: true,
     read_replicas: false,
     prepared_statements: false,
+    isolation_level: false,
     connectionSettings: false,
     retries: false,
+    pool_timeout: false,
+    connection_lifetime: false,
   },
 };