+++ console/src/metadata/sourcesUtils.ts
@@ -1,27 +1,28 @@
 import { Driver } from '../dataSources';
-import { SourceConnectionInfo } from './types';
+import {
+  ConnectionPoolSettings,
+  IsolationLevelOptions,
+  SourceConnectionInfo,
+} from './types';
 
 export const addSource = (
   driver: Driver,
   payload: {
     name: string;
     dbUrl: string | { from_env: string };
-    connection_pool_settings?: {
-      max_connections?: number;
-      idle_timeout?: number;
-      retries?: number;
-    };
+    connection_pool_settings?: ConnectionPoolSettings;
     replace_configuration?: boolean;
     bigQuery: {
       projectId: string;
       datasets: string;
     };
     preparedStatements?: boolean;
+    isolationLevel?: IsolationLevelOptions;
   },
   // supported only for PG sources at the moment
   replicas?: Omit<
     SourceConnectionInfo,
-    'connection_string' | 'use_prepared_statements'
+    'connection_string' | 'use_prepared_statements' | 'isolation_level'
   >[]
 ) => {
   const replace_configuration = payload.replace_configuration ?? false;
@@ -69,6 +70,7 @@ export const addSource = (
           database_url: payload.dbUrl,
           pool_settings: payload.connection_pool_settings,
           use_prepared_statements: payload.preparedStatements,
+          isolation_level: payload.isolationLevel,
         },
         read_replicas: replicas?.length ? replicas : null,
       },